requirements.txt
# ==================== 核心框架 ====================
langchain>=0.1.0
langchain-community>=0.0.20
langchain-classic>=0.1.0
streamlit>=1.31.0

# ==================== 向量数据库 ====================
chromadb>=0.4.22

# ==================== 嵌入和重排序模型 ====================
sentence-transformers>=2.3.0
FlagEmbedding>=1.2.5

# ==================== 大语言模型 ====================
ollama>=0.1.6

# ==================== 文档处理 ====================
pypdf>=3.17.0
python-docx>=1.1.0
markdown>=3.5.0
python-dotenv>=1.0.0

# ==================== 工具库 ====================
numpy>=1.24.0
pandas>=2.0.0
rank-bm25>=0.2.2
tiktoken>=0.5.0
